+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All Notable changes to `Csv` will be documented in this file
 
 - `SwapDelimiter` stream filter to allow working with multibyte CSV delimiter
 - `League\Csv\Serializer\AfterMapping` to work around the limitation aroud constructor usage.
+- `Denormalizer` can register type alias to simplify callback usage.
 
 ### Deprecated